requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.10
langchain==0.3.4
langchain-core==0.3.15
langchain-google-genai==2.0.4
gradio==5.4.0
gradio_client==1.4.2
pydantic==2.9.2
google-generativeai==0.8.3
google-api-python-client==2.151.0
protobuf==5.28.3
